Removed defaults and extracted common cloudformation parameters and helpers
