+++ magenta-lib/src/main/scala/magenta/deployment_type/CloudFormation.scala
@@ -1,13 +1,10 @@
 package magenta.deployment_type
 
 import magenta.artifact.S3Path
+import magenta.deployment_type.CloudFormationDeploymentTypeParameters._
 import magenta.tasks.{CheckUpdateEventsTask, UpdateCloudFormationTask}
-import magenta.tasks.UpdateCloudFormationTask.{LookupByName, LookupByTags}
 
-object CloudFormation extends DeploymentType {
-
-  type TagCriteria = Map[String, String]
-  type CfnParam = String
+object CloudFormation extends DeploymentType with CloudFormationDeploymentTypeParameters {
 
   val name = "cloud-formation"
   def documentation =
@@ -30,21 +27,6 @@ object CloudFormation extends DeploymentType {
       |to this bucket and sent to CloudFormation using the template URL parameter.
     """.stripMargin
 
-  val cloudformationStackByTags = Param[Boolean]("cloudFormationStackByTags",
-    documentation =
-      """When false we derive the stack using the `cloudFormationStackName`, `prependStackToCloudFormationStackName` and
-        |`appendStageToCloudFormationStackName` parameters. When true we find the stack by looking for one with matching
-        |stack, app and stage tags in the same way that autoscaling groups are discovered.""".stripMargin
-  ).defaultFromContext((pkg, _) => Right(!pkg.legacyConfig))
-  val cloudFormationStackName = Param[String]("cloudFormationStackName",
-    documentation = "The name of the CloudFormation stack to update"
-  ).defaultFromContext((pkg, _) => Right(pkg.name))
-  val prependStackToCloudFormationStackName = Param[Boolean]("prependStackToCloudFormationStackName",
-    documentation = "Whether to prepend '`stack`-' to the `cloudFormationStackName`, e.g. MyApp => service-preview-MyApp"
-  ).default(true)
-  val appendStageToCloudFormationStackName = Param[Boolean]("appendStageToCloudFormationStackName",
-    documentation = "Whether to add '-`stage`' to the `cloudFormationStackName`, e.g. MyApp => MyApp-PROD"
-  ).default(true)
   val templatePath = Param[String]("templatePath",
     documentation = "Location of template to use within package. If it has a standard YAML file extension (`.yml` or `.yaml`), the template will be converted from YAML to JSON."
   ).default("""cloud-formation/cfn.json""")
@@ -69,21 +51,6 @@ object CloudFormation extends DeploymentType {
     documentation = "If set to true then the cloudformation stack will be created if it doesn't already exist"
   ).default(true)
 
-
-  val amiTags = Param[Map[String, String]]("amiTags",
-    documentation = "Specify the set of tags to use to find the latest AMI"
-  ).default(Map.empty)
-
-  val amiParameter = Param[String]("amiParameter",
-    documentation = "The CloudFormation parameter name for the AMI"
-  ).default("AMI")
-
-  val amiParametersToTags = Param[Map[CfnParam, TagCriteria]]("amiParametersToTags",
-    documentation =
-      """AMI cloudformation parameter names mapped to the set of tags that should be used to look up an AMI.
-      """.stripMargin
-  ).default(Map.empty)
-
   val updateStack = Action("updateStack",
     """
       |Apply the specified template to a cloudformation stack. This action runs an asynchronous update task and then
@@ -95,28 +62,9 @@ object CloudFormation extends DeploymentType {
       implicit val artifactClient = resources.artifactClient
       val reporter = resources.reporter
 
-      val amiParameterMap: Map[CfnParam, TagCriteria] = (amiParametersToTags.get(pkg), amiTags.get(pkg)) match {
-        case (Some(parametersToTags), Some(tags)) =>
-          reporter.warning("Both amiParametersToTags and amiTags supplied. Ignoring amiTags.")
-          parametersToTags
-        case (Some(parametersToTags), _) => parametersToTags
-        case (None, Some(tags)) => Map(amiParameter(pkg, target, reporter) -> tags)
-        case _ => Map.empty
-      }
+      val amiParameterMap: Map[CfnParam, TagCriteria] = getAmiParameterMap(pkg, target, reporter)
 
-      val cloudFormationStackLookupStrategy = {
-        if (cloudformationStackByTags(pkg, target, reporter)) {
-          LookupByTags(pkg, target, reporter)
-        } else {
-          LookupByName(
-            target.stack,
-            target.parameters.stage,
-            cloudFormationStackName(pkg, target, reporter),
-            prependStack = prependStackToCloudFormationStackName(pkg, target, reporter),
-            appendStage = appendStageToCloudFormationStackName(pkg, target, reporter)
-          )
-        }
-      }
+      val cloudFormationStackLookupStrategy = getCloudFormationStackLookupStrategy(pkg, target, reporter)
 
       val globalParams = templateParameters(pkg, target, reporter)
       val stageParams = templateStageParameters(pkg, target, reporter).lift.apply(target.parameters.stage.name).getOrElse(Map())
